requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
pymongo==4.12.0
qdrant-client==1.14.2
minio==7.2.15
requests==2.32.3
pillow==11.2.1
torch==2.7.0
transformers==4.51.3
numpy==2.2.5
python-multipart==0.0.20
tqdm==4.67.1
